+++ shoko_ui/lib/src/theme/theme.dart
@@ -1,0 +1,6 @@
+library shoko_ui.theme;
+
+export 'theme_widget.dart';
+export 'context_theme_extension.dart';
+export 'styles/radius.dart';
+export 'styles/shadow.dart';