+++ shoko_ui/lib/src/widgets/code/code_widget.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:shoko_ui/shoko_ui.dart';
+import 'package:shoko_ui/src/theme/styles/radius.dart';
 
 enum ShokoUICodeBorderType {externalOutline, underSymbolsBorder}
 
@@ -28,7 +29,7 @@ class ShokoCodeTextField extends StatefulWidget {
   final bool autoNext;
   final bool fillBackground;
 
-  final ShokoRadii? radius;
+  final SRadii? radius;
 
   final double? height;
   final double? width;
@@ -119,7 +120,7 @@ class _ShokoCodeTextFieldState extends State<ShokoCodeTextField> {
     return AnimatedContainer(
       duration: const Duration(seconds: 1),
       decoration: widget.borderType == ShokoUICodeBorderType.externalOutline ? BoxDecoration(
-        borderRadius: widget.radius?.get() ?? ShokoRadii.medium.get(),
+        borderRadius: widget.radius?.get() ?? SRadii.medium.get(),
         border: Border.all(
           color: widget.enableColor ?? Colors.black
         )
@@ -170,7 +171,7 @@ class _CodeWidgetSymbol extends StatelessWidget {
   final bool fillBackground;
 
   final Color color;
-  final ShokoRadii? radius;
+  final SRadii? radius;
 
   final double height;
   final double width;
@@ -190,7 +191,7 @@ class _CodeWidgetSymbol extends StatelessWidget {
   BoxDecoration? _getDecoration() {
     if (fillBackground) {
       return BoxDecoration(
-        borderRadius: radius?.get() ?? ShokoRadii.medium.get(),
+        borderRadius: radius?.get() ?? SRadii.medium.get(),
         color: color
       );
     }